+++ lib/resources.js
