+++ lib/routes.js
@@ -1,6 +1,6 @@
 var Boom = require('boom');
 var Community = require('./community');
-var Learn = require('./learn');
+var Resources = require('./resources');
 var Markdown = require('./markdown');
 var Path = require('path');
 var Plugins = require('./plugins');
@@ -229,11 +229,11 @@ exports.routes.push({
 
 exports.routes.push({
     method: 'GET',
-    path: '/learn',
+    path: '/resources',
     config: {
         handler: function (request, reply) {
 
-            reply.view('learn', { learn: Learn });
+            reply.view('resources', { resources: Resources });
         }
     }
 });
