+++ templates/includes/footer.jade
@@ -14,7 +14,7 @@ footer(role='contentinfo')
         li
           a(href='/plugins', title='Plugins') Plugins
         li
-          a(href='/learn', title='Learn') Learn
+          a(href='/resources', title='Resources') Resources
         li
           a(href='https://github.com/hapijs/hapi', title='hapi on Github') Github
         li
