+++ templates/includes/navigation.jade
@@ -18,6 +18,6 @@ nav.animated.fadeInDown(role='navigation')
       li
         a(href='/contribute', title='Contribute to hapi') Contribute
       li  
-        a(href='/learn', title='Learn') Learn
+        a(href='/resources', title='Resources') Resources
       li
         a(href='/help') Need Help?
