--- templates/learn.jade
@@ -1,30 +1,0 @@
-extends layout
-
-block title
-  title Learn hapi
-
-block content
-  .content.page-api.cf
-
-    .column.api-description
-      p Aside from the content that you'll find on this site, there's also a lot of community-produced content that can help you learn how to be effective with hapi.js. If you've produced some content that you want to see on this list, please send us a <a href="https://github.com/hapijs/hapijs.com/blob/beta/lib/learn.js">pull request</a>. 
-
-      each category, categoryName in learn.categories
-        .category
-          a(id=categoryName)
-          h3= categoryName
-
-          ul
-            each item, itemName in category
-              li
-                a(href=item.url)= itemName
-                p= item.description
-
-    .column.nav-secondary
-      nav(role='toc')
-        h3 Learn
-
-        ul.unstyled(role='categories')
-          each category, categoryName in learn.categories
-            li(role='category')
-              a(href='##{categoryName}')= categoryName
