+++ templates/resources.jade
@@ -1,0 +1,30 @@
+extends layout
+
+block title
+  title hapi Resources
+
+block content
+  .content.page-api.cf
+
+    .column.api-description
+      p Aside from the content that you'll find on this site, there's also a lot of community-produced content that can help you learn how to be effective with hapi.js. If you've produced some content that you want to see on this list, please send us a <a href="https://github.com/hapijs/hapijs.com/blob/beta/lib/resources.js">pull request</a>. 
+
+      each category, categoryName in resources.categories
+        .category
+          a(id=categoryName)
+          h3= categoryName
+
+          ul
+            each item, itemName in category
+              li
+                a(href=item.url)= itemName
+                p= item.description
+
+    .column.nav-secondary
+      nav(role='toc')
+        h3 Resources
+
+        ul.unstyled(role='categories')
+          each category, categoryName in resources.categories
+            li(role='category')
+              a(href='##{categoryName}')= categoryName
